Moved blank application to a separated file
