+++ shared/ui-components/index.ts
@@ -59,6 +59,7 @@ export * from "./src/helpers/useLanguageChange"
 export * from "./src/helpers/debounce"
 export * from "./src/helpers/formatIncome"
 export * from "./src/helpers/validators"
+export * from "./src/helpers/blankApplication"
 /* Icons */
 export * from "./src/icons/HeaderBadge"
 export * from "./src/icons/Icon"
